Resolve resource deadlock in command monitor (#7110)
Prevent command monitor thread from joining itself.

Fallout from #7100 when a command monitor thread invokes command
completion notification.  The command callback functon takes a
reference to the command object itself.  This object in turn owns a
reference to the hwqueue that manages command execution.  If this is
the last reference to the command and the last reference to the
hwqueue, then the hwqueue is deleted on the monitor thread itself.

Before this commit, the hwqueue owns the command monitor thread and
would join with the thread upon destruction.  This results in resource
deadlock execption when the current thread is the same as the monitor
thread.

With this commit, the command manager (and the associated monitor
thread) is managed indepdently of the hwqueue implementation and
disposed command managers are recycled for later use but only
destructed at process exit.