fix typo in ert_cmd_state comment
This change corrects the commenting on the ert_cmd_state enum. Previously, the comment for ERT_CMD_STATE_COMPLETED listed the value as ERT_CMD_STATE_COMPLETE. This reads as a reasonable name, but ended up wasting a good 25 minutes of my time. Now, the commenting (and implicitly any docs generated from it) will correctly inform users of the correct possible values of the enum