Extend xrt::bo implementation with hardware context
Add new internal class to capture both core device and hardware
context as bo_impl data members.

This PR add the basics infrastructure for explicitly associating a
xrt::bo object with a hardware context rather than embedded a context
handle in bo flags.  This also implies that a bo_impl object owns a
reference to the hardware context in which it is created thus ensuring
that BOs are deleted before the hwctx is deleted.

For legacy reasons, the public xrt::bo constructors use
xclDeviceHandle when they really should have used xrt::device objects.
This PR is a stepping stone for follow up work to deprecate the
xrt::bo constructors that use xclDeviceHandle in favor of new variants
that take xrt::hw_context.