Reinstance xrt::bo constructors from device object
Allocating an xrt::bo for device was premature deprecated in PR #7165,
this PR instances the constructors but this time with xrt::device
objects rather than xclDeviceHandle.

The original xclDeviceHandle variants remain deprecated and will be
removed in a future release.

In #7165 constructors specific to a hardware context were added.
These contructors remain for the cases where buffers truly are
hardware context specific.  The implmentation is free to ignore the
hwctx if it is not applicable to the type of buffer being created,
e.g. for cases where the buffer can in fact be shared accross device
process and is not specific to a hwctx.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>