Modify xrt_hwctx_handle to support all the variants of a hwctx
This PR changes xrt_hwctx_handle into a class that supports holding
either an class derived from xrt_hwctx_handle, a void*, or a slotidx.

This change is necessary for shims that migrate towards representing
a hwctx as an opaque handle. By converting xrt_hwctx_handle into a base
class, a shim derived class can redefine methods for accessing private
details, e.g. slotidx, of a context handle.

The slotidx is currently used by xrt::kernel::group_id to encode in
flags used when constructing an xrt::bo.  A new hwctx aware xrt::bo
constructor has been added, but legacy applications do not use this
constructor and continue to rely on the encoding of slotidx.

Until the encoding can be removed, the class xrt_hwctx_handle has a
virtual override that can be redefined by shim to return the slotidx,
this allows the existing encoding to continue to be used.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>