Add shim handle for hardware context and queue
This PR enforces all shim layers to represent a hardware context
(and queue) as handles derived off of shim base classes.

The base classes are used to represent shim objects to XRT coreutil
and provide an interface that use to be exposed at the
xrt_core::device level.

The PR is a small step towards a general rearchitecture of shim level
where all shim objects are represented as objects of classes derived
off of shim handles.  This will simplify and unify shim
implementations and in turn make xrt_core simpler.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>