WIP: Support hardware queue dependencies
- Add xrt_core::fence and xrt_core::fence_handle internal classes
- Add xrt::run::enqueue() with and without dependencies

New methods xrt::run::enqueue() allows a run object to be enqueued
with a dependency on other run objects.  A dependency on a run object
is possible only if that run object has itself been enqueued().  In
other words, legacy xrt::run::start() has not been changed and a run
object started this way cannot be used as dependecy on another run
object submission.