Fix bug in xclAPI implementation of buffer_handle release
Support detatching underlying xclBufferHandle (and export handle) from
managed buffer_handle and shared_handle.

Prior to this change the managed handles were released and the under
xcl handles returned, but this lead to a memory leak because the
managed handle was never deleted.

This commit instead add a detach_handle member function to shim
implementation classes of buffer_handle and shared_handle.  The
managed handles are no longer released but if detach_handle was
called, the destruction of the managed handle doesn't close the
detached handle.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>