Use std::shared_ptr for bo_impl::handle
For most use cases bo_impl is the exclusive owner of a
xrt_core::buffer_handle.  But in a few cases, the buffer_handle must
be shared by multiple bo_impls.

For example, a sub-buffer shares the handle with its parent, and the
handle should be disposed of only when the last reference goes away.

Another example is nodma buffers which is a specialized bo_impl that
contains a device_only and a host_only buffer.  The bo_impl itself
shared the handle with the device_only bo_impl.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>