CR-1158312 XRT regression : xclRegRW: can't map CU (#7470)
Make sure dynamic update of hwctx accces_mode is reflected in shim
specialization that care.

This PR is specific for design that use mailbox. XRT under the hood
changes CU access mode to exclusive because mailbox require register
read/write which is only supported for exclusive mode CUs.  This
update of access_mode was not being reflected in the shim
specialization of hwctx_handle.  The fix adds a new optional virtual
function to hwctx_handle that is used to reflect the access_mode
change.

The bug was introduced when the hwctx_handle was added #7280.