Buffer object experimental extension to support access mode (#7517)
* Fix bug when clearing bo flags

* Experimental simplified extension to construct buffer objects

Provide simpler xrt::bo constructors that support access mode.

Some AIE platforms support only a limited set of buffer types. This
extension provides a simpler bo construction interface that supports
allocating host_only buffers exclusively.

In addition the extension allows bo construction to specify how the
buffer object can be access.  Support is added for sharing between
devices within process or sharing between processes.  At least on some
implmentation, the access decision must be made when the bo is first
constructed.

This extension is experimental for now.

* Support importing a buffer object for use within a hardware context

When an argument bo (arg_bo) is created it is gpu_va is mapped and the
buffer is made resident which requires waiting on a paging operation to
complete before the buffer can be used within in a command within a hwctx.

For this reason, imported bo is construted with a hwctx argument, it
doesn't mean that the bo is specific to this context, but it does mean
that the bo is synchrononized with operations within that context.