Vitis-7921 First-class execution buffer (#7572)
* Vitis-7921 First-class execution buffer

Add first-class concept to represent instructions used by a function
to execute on some configured hardware.

In some scenarios, an xclbin configures the hardware with an
overlay that is used to execute control code / instructions
specific to a function, or put another way, the function executes
instructions on the overlay.

Without a first-class instruction buffer, the host code is passing
instructions as a semi-random regular kernel argument under the
assumption that firmware can locate the instructions from the
arguments.

This PR introduces `xrt::module` for repsenting the a binary that
contains instructions for one or more functions.  The module is used
to create instruction buffers for specific functions and kernel
execution will populate an ERT packet for a function with the
instruction buffer address explicitly defined.

The ERT packet for kernel execution is the standard
`ert_start_kernel_cmd`, but the data payload is reinterpreted as a
`ert_dpu_data` struct, which has explicit fields for instruction
buffer address and size.

Additional thie PR add xrt::elf for managing an elf binary that in
turn can be used to create an xrt::module.  The xrt::elf will
eventually provide APIs for extracting elf meta-data, like function
names, arguments and more.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Support latest revision of xrt::module

Per discussions the flow is changed slightly.

- Construct xrt::module from elf or userptr.
- Construct xrt::kernel from hwctx and module
- Copy module to sram when creating xrt::run from kernel
- Prep for patching of global argument buffer address in xrt::run::set_arg

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Change ert_dpu_data to pass size of buffer in bytes

Instruction count (number of instructions) is to be handled by KMD.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>