Add xrt::ext::bo support for specifying read/write direction (#7607)
Amend xrt::ext::bo::access_mode with `read` and `write` enumerators.
Treat mode as bitflags.

A buffer can be opened for read, meaning the device will read the
content written by host, or it can be opened for write, meaning
the device will write to the buffer and the host will read. To
specify that a buffer is used for both read and write, the access
flags can be ORed.  If neither read or write is specified, the
default is read|write.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>