Amend #7572 to validate ERT_START_DPU opcode
Fix xrt::kernel implementation to amend the size of the command
packet for dpu style kernels that use xrt::module for instructions.

The instructions (when used) are prepended to the command payload and
the size of the payload has adjusted for struct ert_dpu_data.

Update xrt::xclbin to expose the type of the xrt::xclbin::kernel and
showcase in 56_xclbin testcase.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>