Rework ert_dpu_data embedding in command packet
Use proper aligned size of ert_dpu_data when computing command payload
count.

Drop the data[] payload of ert_dpu_data, it is the responsibility of
client code to offset the command packet data payload past
ert_dpu_data if necessary.

Move hwctx association of module to kernel_impl constructor as opposed
to run_impl constructor.  The hwctx is kernel specific, so is the
instruction buffer.  If a different set of instructions are to be used
with a kernel, then a new kernel object should be constructed from a
new module.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>