Amend #7572 to validate ERT_START_DPU opcode (#7628)
* Amend #7572 to validate ERT_START_DPU opcode

Fix xrt::kernel implementation to amend the size of the command
packet for dpu style kernels that use xrt::module for instructions.

The instructions (when used) are prepended to the command payload and
the size of the payload has adjusted for struct ert_dpu_data.

Update xrt::xclbin to expose the type of the xrt::xclbin::kernel and
showcase in 56_xclbin testcase.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Rework ert_dpu_data embedding in command packet

Use proper aligned size of ert_dpu_data when computing command payload
count.

Drop the data[] payload of ert_dpu_data, it is the responsibility of
client code to offset the command packet data payload past
ert_dpu_data if necessary.

Move hwctx association of module to kernel_impl constructor as opposed
to run_impl constructor.  The hwctx is kernel specific, so is the
instruction buffer.  If a different set of instructions are to be used
with a kernel, then a new kernel object should be constructed from a
new module.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>