CR-1167063 Fix hwqueue_handle::wait_command() API comment (#7609)
* CR-1167063 Fix hwqueue_handle::wait_command() API comment

Remove comment about cmd nullptr waiting for any command to complete.

Throw if nullptr is passed.  This means that OpenCL is not supported
with platforms that implement hwqueue. If OpenCL support is needed,
the command monitor needs to track order of submitted commands and
wait on first submitted command before checking status of all running
commands.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* fix windows build error

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>