Vitis-8819  Mine ELF for dynamic relocation info and patch ctrlcode
Parse ELF for dynamic relocation sections and create argument patcher
objects for patching of control code when arguments are known.

The parsing and creation of dynamic relocation data and the creation
of argument address patchers is done at ELF load time.

The xrt::module ELF loader extracts AIE column ctrlcode from the ELF
file and creates the argument patchers from dynamic relocation
sections.

The xrt::module is used when construction an xrt::kernel object. The
kernel's xrt::module serves as a template for xrt::run objects created
from the kernel.  A run object copies the module control code into an
xrt::bo buffer object.  The address of the buffer object is used by
the argument patchers when the run object patches the control code
with buffer objects and scalar argument values.

The underlying assumption is that all column ctrlcode is stored in a
contiguous buffer object at the offsets calculated by the xrt::module
ELF loader.

The code added in this pull request is work-in-progress and very
specific to our current control code (ELF) definition.  As our usage
evolve, the code will be refactored into shim specific implemenations.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>