CR-1172341 XRT Test Code fails with 202320.2.16.175 packages
Move caching of command header to the very first start() of a run object.

In #7608 the internally initialized ert command header is cached as
part of a run object.  The caching is done for repeated invocation of
same run object to reset the header to a known good initial state.
However, caching during construction of the run object is to early for
some use cases where the header is changed in the host code to support
legacy command opcodes.

This PR fixes the problem by delaying caching until the very first
invocation of xrt::run::start().  This allows for host code to
manipulate the underlying command packet, which is necessary for
support of legacy ERT_EXEC_WRITE.