Work-around for CI build failure on versal
The version.h is generated at CMake time from
XRT/src/CMake/config/version.h.in by XRT/src/CMake/version.cmake.

version.cmake uses ${GIT_EXECUTABLE} to get meta data specific to
version of XRT being built, but none of the data is gathered in CI
when running `build_edge.sh -aarch versal -full` and the version.h
file is missing Git data which instead ends up as empty string.

XRT_BUILD is a define macro based on number of commits acquired from
Git, but if this value is empty the use of the macro fails.  This
PR defaults the value to 0 in hopes it gets overwritten by Git.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>