Support concept of feature id
The feature id (xrt::version::feature()) is defined as the total
number of commits to XRT main branch.  For branches off of XRT's main
branch, the feature number is the total number of commits at the time
the branch diverged from XRT's main branch.

The feature id is computed as the difference between number of commits
on current branch and the number of commits since origin of branch from
master.

(git rev-list HEAD --count) - (git rev-list HEAD ^origin/master --count)

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>