Add build.sh option to force error if ERT FW cannot be build (#7752)
When the build is initiated by CI system, it is an error to build XRT
without MicroBlaze firmware. The FW is needed when platform packages
are installed in order to create xsabin.

This PR add `-ci` build option to be set by CI when building XRT on
Linux.  The option is used to error out if ERT FW cannot be built.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>