Amend #7773 to add back missing include (#7793)
Add back include of gen/version.h to core/common/system.cpp.

The #include of gen/version.h was removed from system.cpp in #7773,
but this header is used by system::get_xrt_build_info().  At least on
my system it breaks the build. I really bumped it passed the pipeline,
and mystified why it is failing to build on my ubuntu20.04 system. I
am adding it back in.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>