EDGEML-6881 Random test failure when looking for driver store path
Adjust copying of wchar[] array into std::wstring accounting for
bytes vs wchar.

Code queries windows adpater for driver store path.  The driver store path is
returned as array of wchar but was copied as bytes, resulting in copying from
unallocated memory.

This bug was introduced in #7679 and is present in MCDM drops starting M6.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>