CR-1181687 Fix hwctx_handle::create_hw_queue() to return shared queue
Share the ownership of xrt_core::hwqueue_handle with shim.

A hwqueue is managed by hwctx (today 1:1), and used by coreutil XRT
(coreutil) for submitting commands for execution and more. XRT shim
may itself use the hwqueue for miscellaneous commands and as such the
underlying queue is actually owned by shim (hwctx_handle) but shared
with coreutil.

This PR makes it explicit that the hwqueue ownership is with the
hwctx_handle. XRT coreutil no longer creates a hwqueue but instead
gets a hwqueue_handle from the hwctx_handle.

Prior to this change, there were strange hacks at shim level to return a
unique_ptr to something that is actually shared.  This strange logic
can now be changed.

It is the responsibility of XRT coreutil to ensure that lifetime of a
hwctx extends the use of the hwqueue_handle that was obtained from the
hwctx_handle.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>