Add support for debug buffer object in XRT for internal use
Allocate a “special” type of BO in a virtual region shared with
instruction buffers within a hardware context.

The driver will share this buffer object with firmware where the
buffer will be populated with data to be shared with user space XRT.

The implementation makes use of a new XRT_BO_USE_DEBUG flag which is
ignored for all other use cases.

The allocation of the debug buffer object is through an internal API
which takes as argument the xrt::hw_context in which the buffer is to
be allocated, plus the size of the debug buffer.