Include GSL as a submodule
Add GSL as submodule in src/runtime_src/core/common/gsl.

This library is MIT licensed and provides some features that are yet
to be included in std:: for some of the compilers we use.

In particular GSL provides gsl::span, which we will alias to
xrt_core::span.  The idea is that xrt::span will be re-aliased
to std::span alias when supported.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>