VITIS-11806 Break runlist into multiple submissions if necessary
Break an xrt::runlist into multiple submissions of max fixed size.

Device supports only fixed number of chained commands, so logic is
moved to XRT coreutil to break a runlist into multiple chunks.

The internal hwqueue API for runlist submission is changed to take a
std::span of buffer_handles rather than a vector.  This allows for
easier / cheaper submission as it avoids creating a new vector
everytime a chunk is submitted.

To leverage std::span on compilers that do not yet support this, XRT
now includes Microsoft GSL (Guidelines Support Library) as a
submodule.  This library comes with gsl::span which we use until
std::span is supported by all compilers (c++20).

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>