VITIS-11806 Break runlist into multiple submissions if necessary (#8171)
* Include GSL as a submodule

Add GSL as submodule in src/runtime_src/core/common/gsl.

This library is MIT licensed and provides some features that are yet
to be included in std:: for some of the compilers we use.

In particular GSL provides gsl::span, which we will alias to
xrt_core::span.  The idea is that xrt::span will be re-aliased
to std::span alias when supported.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* VITIS-11806 Break runlist into multiple submissions if necessary

Break an xrt::runlist into multiple submissions of max fixed size.

Device supports only fixed number of chained commands, so logic is
moved to XRT coreutil to break a runlist into multiple chunks.

The internal hwqueue API for runlist submission is changed to take a
std::span of buffer_handles rather than a vector.  This allows for
easier / cheaper submission as it avoids creating a new vector
everytime a chunk is submitted.

To leverage std::span on compilers that do not yet support this, XRT
now includes Microsoft GSL (Guidelines Support Library) as a
submodule.  This library comes with gsl::span which we use until
std::span is supported by all compilers (c++20).

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix windows compile warnings

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Update comments

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Limit runlist size to 24 per submission

This is very arbitrary and TBD.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix bugs

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Update NOTICE for including Microsoft GSL

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>