Fixed ERT firmware packaging issue
When proving the user's own sched*.bin binaries, the ERTFW_FILES variable contains the required ERT firmware files but it is only used in one target firmware location in the target install directory structure. This fix ensure it is also added to the directory expected by the xrt base packages.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>