VITIS-11806 Support runlist submission as chained command objects
Add new ERT_CMD_CHAIN opcode for chained command submssion of
a runlist.  A runlist is broken into multiple chained commands if
necessary.

This PR reworks #8171 and replaces disjoint set submission of individual
command buffers with submission of a single ert_packet command buffer
that chains individual commands.

The chained command is represented as an `ert_packet` where the payload is
interpreted as `ert_cmd_chain_data` per ert.h.  The pay holds an array
of buffer_handles that are to be submitted atomically for execution.

This chained command submission is somewhat easier to manage and is a
step towards supporting command fusing where multiple run objects are
stiched together into a single command without managing individual
run objects.

This PR also parameterizes bo_cache with the size of bos needed. This
is done because runlist will not currently need a 4K buffer for the
chained ert packet.  This may change in future when runlist fuses the
individual commands.  It is understood that 4K may in fact be the
minimum underlying allocation size, but that is unbeknowst to the code
that uses the bo cache.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>