Redo error handling if bind_at throws
There is one caveat here and that is the chain_data command_count is
one short of what is added to the array.  This is to avoid state
change if bind_at throws, but is of course no good if bind_at expects
argument `idx` to match up with `command_count`.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>