Support polling for runlist completion (#8253)
Per side band request implement polling for xrt::runlist completion
as well as polling for xrt::run completion.

This change also closes a minor bug in xrt::run::state() which
was not supported properly on Windows (MCDM) where command state is
lazy updated only as part of calling xrt::run::wait().

xrt::run::state() is implemented to invoke polling which is a noop on
Linux where the state is updated automatically.  On Windows polling
will check the fence associated with the command and complete the
command if the fence is signaled.

In addition this PR implements xrt::runlist::poll() which polls for
completion of the last chained command submitted for execution. As for
get_state(), polling either returns immediately if command is still
running, or completes/updats the state of the command if it has
finished.

There is no explicit poll() function for xrt::run, here
xrt::run::state() is polling for the run state.

A new API xrt::runlist::poll() was added to poll for runlist
completion.

Testing was manual in MCDM.

There is no risk here, as functionality doesn't intersect existing
APIs.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>