linux/xocl: Adjust to platform_driver::remove() return void
See the upstream linux commit 0edb555a65d1 ("platform: Make
platform_driver::remove() return void").

Generated with the following Coccinelle script:
> @rule_remove@
> identifier D;
> identifier func;
> @@
> static struct platform_driver D = {
> ...,
>         .remove         = func,
> ...
> };
> @rule_func depends on rule_remove@
> identifier pdev;
> identifier rule_remove.func;
> fresh identifier Fn = "__" ## func;
> @@
> static int
> - func(struct platform_device *pdev)
> + Fn(struct platform_device *pdev)
> {
> ...
> }
> +
> + #if LINUX_VERSION_CODE >= KERNEL_VERSION(6, 11, 0)
> + void func(struct platform_device *pdev)
> + {
> +        Fn(pdev);
> + }
> + #else
> + #define func Fn
> + #endif

Though, it added new function with a brace on the same line, so I ran
checkpath.pl --fix on the resulting patch.
And then it seems there is either bug in spatch or it doesn't know
how to deal with 'static', which prevented me from adding
'static void func(...)', so on top of checkpatch.pl, I ran
> sed 's/^+\(void [[:print:]]\+_remove(struct platform_device \*pdev)\)$/+static \1/'

And yet the last bit is also manual: I had to remove a stray semicolon,
that was living after ps_remove() last curly brace, which post-spatch
became '#endif;' and the compiler didn't quite liked it.

Signed-off-by: Dmitry Safonov <dima@arista.com>