CR-1217444 Callback on xrt::run causes "kds monitor died" and hang
Trap unsupposed managed execution earlier.

We don't plan on supporting callbacks for xrt::run objects for
RyzenAI.  This would kill performance.  Callbacks are used under the
hood in the OpenCL implementation and are indeed supported for XRT
native APIs on Alveo, but not for RyzenAI

This PR changes code such that an exception is thrown when an xrt::run
object is started after adding a callback on the run object.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>