Support windows linking with xrt_coreutil_static without dllimport
Add a preprocessor directive to turn off dllimport decoration of
exported symbols when exported APIs are intended to be statically
linked into some external target.

On windows LNK4217 means that a symbol defined in one object (e.g., a
static link library) is being imported by another object.  In order to
avoid this situation, the __declspec(dllimport) modifier should be
removed from the object that uses the symbol.

This PR simply adds XRT_STATIC_BUILD preprocessor directive that can
be used when compiling objects that use XRT symbols and link to static
libraries from XRT.

The changes in this PR are target for xrt_coreutil_static usage only,
specifically for use in a user-mode-driver for DirectML.