WIP: Add runner for run-recipe
A run-recipe defines how to execute a graph model using XRT.

This changes contain a stand-alone runner class that reads and
executes a run-recipe json file. The idea is to have tools
geneate the run-recipe along with xclbin and control code for kernels.

The format (schema) of the recipe json is loosely defined. The
implementation of the runner drove some of the defintion of the json
format.

A run-recipe is associated with exactly one xclbin which, when loaded
into a region (partition) on the device, can run the recipe.

This is work-in-progress and is base-lined and validated for today's
NPU TXN control code.  The runner will change shortly when we obsolete
xclbin for NPU.

The recipe will not work with Alveo as the runner uses xrt::runlist
for execution and xrt::runlist is not supposed in Alveo.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>