cppcheck - various fixes
Add nullptr check; not that this will ever trigger, but sure, cppcheck
is correct about missing nullptr check.

Use override on virtual function overrides, but leave virtual if the
override makes the function pure virtual for additional derived classes.

Use explicit to avoid converting constructors, but allow converting
constructors for classes that act more or less as alias for type.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>