Fix xrt-smi throughput tests to measure throughput
Fix Throughput tests for xrt-smi to actually measure throughput.

NPU test for chained command throughput was measuring latency but
computing throughput.  Fix to measure through ensure NPU is always
busy.  This was done by using two xrt::runlists.

Additional problem for chained command througput test was the
iteration count, which was too small to hide the ramp-up cost of clock
scaling, so the test was fixed to increase the iteration count and
lowering the runlist command count.

NPU test for individual command throughput was missing wait on last
batch of commands submitted.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>