Update aie-rt submodule to be in sync with aiebu/lib/aie-rt
This PR updates the aie-rt submodule at src/runtime_src/aie-rt to be
in synch with what aiebu submodule expects.  The aiebu submodule is
updated in earlier commits of this PR.

This is a mess.  Why two copies of aie-rt?  But nevermind, on windows
aiebu is using src/runtime_src/aie-rt and this version of aie-rt needs
to match what aiebu is expecting.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>