Support AIE columns with multiple controllers
When extracting MicroBlaze control code from ELF for a multi-column
partition, account for each column having multiple controllers and
that a partition may not use all of these.

ELF sections identify which column controllers are used.  Currently
this done through an ELF section naming convention.

Before this change, each column had exactly one controller and it was
assumed that either the lead controller in a partition was used alone
or with all remaining controllers in subsequent columns. With this PR,
support is added for multiple controllers in each column and for only
a subset of these controllers being used.

The ERT packet payload interpretation for ert_dpu_data is changed to
indicate which column processor the data is associated with.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>