Build XDP depending on XRT component being built
Change how XDP is built in XRT.  Infer XDP build options based
on which component is being built.

XDP is included in the npu and alveo components and the in the legacy
xrt component, but it is being built differently depending on the
component. XDP is excluded from XRT's base component.

This PR infers compile flags for XDP based on the component.
XDP_CLIENT_BUILD_CMAKE is now an internal configuration variable for
XDP.

There are no visible changes in this PR.  However, anything invoking
CMake with XDP_CLIENT_BUILD_CMAKE will be affected and must change to
ensure that XRT is building the NPU component by adding `-DXRT_NPU`
when invoking CMake.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>