Create XRT Windows SDK when base-dev component is built
When building the base-dev component on windows, support creating
an SDK installer with the content of the development package.

A new build option `build22.bat -sdk` enables building of the SDK,
but requires NSIS to be installed on the build host.

The `-sdk` build switch implies building for npu, which
in turn builds both the base and npu components of XRT.  In order to
prevent build failures for folks who have not installed NSIS, the sdk
is only built when explicitly requested through `-sdk`.

```
build22.bat -npu

build22.bat -sdk

build22.bat -npu -sdk
```

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>