Add CMAKE_INSTALL_PREFIX when building on WSL
Set CMAKE_INSTALL_PREFIX in build script to default or specified
location.

In #8758, the control of CMAKE_INSTALL_PREFIX was removed from
nativeWin.cmake, but this results in default install prefix is
`C:\Program Files` if not explicitly set by build script.

In this PR, the WSL build script is amended to use default install
prefix if none is specified.  The default behavior mirrors that of
build22.bat.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>