Remove default arguments from xrt::hw_context constructor
The decision about the value of unused arguments is moved
to the implementation code for ABI protection.

This PR additionally add shim callback for creating a hwctx_handle
from an elf.  The default implementation (if not overridden by shim)
uses existing callback that takes the partition size.

xrt::elf was enhanced to extract the parittion size from the elf
without going through a module.

xrt::module is in dire need of a rewrite.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>