Extend runner with an execution profile
This WIP.   Adding execution profile data that constraints how
a run recipe is executed by the xrt::runner class.

The profile is a separate json that is paired with a recipe when
instantiating an xrt::runner.

The profile is optional, while the recipe is required.  The
application can use the xrt::runner APIs to bind external resources or
use the profile json to specify the binding and otherwise constaint
the execution of a recipe.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>